Move Bugzilla client to new jbi.bugzilla.client module
