+++ tests/unit/bugzilla/test_client.py
@@ -2,9 +2,8 @@ import pytest
 import responses
 from responses import matchers
 
-from jbi.environment import get_settings
+from jbi.bugzilla.client import BugzillaClient, BugzillaClientError
 from jbi.models import BugzillaWebhookRequest
-from jbi.services.bugzilla import BugzillaClientError, get_service
 
 
 @pytest.fixture
@@ -21,12 +20,20 @@ def webhook_private_comment_example(
     return webhook_payload
 
 
+@pytest.fixture
+def bugzilla_client(settings):
+    return BugzillaClient(
+        base_url=settings.bugzilla_base_url, api_key=settings.bugzilla_api_key
+    )
+
+
 @pytest.mark.no_mocked_bugzilla
-def test_timer_is_used_on_bugzilla_get_comments(mocked_responses, mocked_statsd):
-    bugzilla_client = get_service().client
+def test_timer_is_used_on_bugzilla_get_comments(
+    bugzilla_client, settings, mocked_responses, mocked_statsd
+):
     mocked_responses.add(
         "GET",
-        f"{get_settings().bugzilla_base_url}/rest/bug/42/comment",
+        f"{settings.bugzilla_base_url}/rest/bug/42/comment",
         json={
             "bugs": {"42": {"comments": []}},
         },
@@ -36,8 +43,10 @@ def test_timer_is_used_on_bugzilla_get_comments(mocked_responses, mocked_statsd)
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_methods_are_retried_if_raising(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/bug/42/comment"
+def test_bugzilla_methods_are_retried_if_raising(
+    bugzilla_client, settings, mocked_responses
+):
+    url = f"{settings.bugzilla_base_url}/rest/bug/42/comment"
     mocked_responses.add(responses.GET, url, status=503, json={})
     mocked_responses.add(
         responses.GET,
@@ -48,14 +57,14 @@ def test_bugzilla_methods_are_retried_if_raising(mocked_responses):
     )
 
     # Not raising
-    get_service().client.get_comments(42)
+    bugzilla_client.get_comments(42)
 
     assert len(mocked_responses.calls) == 2
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_key_is_passed_in_header(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/whoami"
+def test_bugzilla_key_is_passed_in_header(bugzilla_client, settings, mocked_responses):
+    url = f"{settings.bugzilla_base_url}/rest/whoami"
     mocked_responses.add(
         responses.GET,
         url,
@@ -65,7 +74,7 @@ def test_bugzilla_key_is_passed_in_header(mocked_responses):
         ],
     )
 
-    assert get_service().client.logged_in()
+    assert bugzilla_client.logged_in()
 
     assert len(mocked_responses.calls) == 1
     # The following assertion is redundant with matchers but also more explicit.
@@ -73,46 +82,52 @@ def test_bugzilla_key_is_passed_in_header(mocked_responses):
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_raises_if_response_has_error(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/bug/42"
+def test_bugzilla_raises_if_response_has_error(
+    bugzilla_client, settings, mocked_responses
+):
+    url = f"{settings.bugzilla_base_url}/rest/bug/42"
     mocked_responses.add(
         responses.GET, url, json={"error": True, "message": "not happy"}
     )
 
     with pytest.raises(BugzillaClientError) as exc:
-        get_service().client.get_bug(42)
+        bugzilla_client.get_bug(42)
 
     assert "not happy" in str(exc)
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_get_bug_raises_if_response_has_no_bugs(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/bug/42"
+def test_bugzilla_get_bug_raises_if_response_has_no_bugs(
+    bugzilla_client, settings, mocked_responses
+):
+    url = f"{settings.bugzilla_base_url}/rest/bug/42"
     mocked_responses.add(responses.GET, url, json={"bugs": []})
 
     with pytest.raises(BugzillaClientError) as exc:
-        get_service().client.get_bug(42)
+        bugzilla_client.get_bug(42)
 
     assert "Unexpected response" in str(exc)
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_get_comments_raises_if_response_has_no_bugs(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/bug/42/comment"
+def test_bugzilla_get_comments_raises_if_response_has_no_bugs(
+    bugzilla_client, settings, mocked_responses
+):
+    url = f"{settings.bugzilla_base_url}/rest/bug/42/comment"
     mocked_responses.add(responses.GET, url, json={"bugs": {"42": {}}})
 
     with pytest.raises(BugzillaClientError) as exc:
-        get_service().client.get_comments(42)
+        bugzilla_client.get_comments(42)
 
     assert "Unexpected response" in str(exc)
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_update_bug_uses_a_put(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/bug/42"
+def test_bugzilla_update_bug_uses_a_put(bugzilla_client, settings, mocked_responses):
+    url = f"{settings.bugzilla_base_url}/rest/bug/42"
     mocked_responses.add(responses.PUT, url, json={"bugs": [{"id": 42}]})
 
-    get_service().client.update_bug(42, see_also={"add": ["http://url.com"]})
+    bugzilla_client.update_bug(42, see_also={"add": ["http://url.com"]})
 
     assert (
         mocked_responses.calls[0].request.body
@@ -121,10 +136,12 @@ def test_bugzilla_update_bug_uses_a_put(mocked_responses):
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_get_bug_comment(mocked_responses, webhook_private_comment_example):
+def test_bugzilla_get_bug_comment(
+    bugzilla_client, settings, mocked_responses, webhook_private_comment_example
+):
     # given
     bug_url = (
-        f"{get_settings().bugzilla_base_url}/rest/bug/%s"
+        f"{settings.bugzilla_base_url}/rest/bug/%s"
         % webhook_private_comment_example.bug.id
     )
     mocked_responses.add(
@@ -164,7 +181,7 @@ def test_bugzilla_get_bug_comment(mocked_responses, webhook_private_comment_exam
         },
     )
 
-    expanded = get_service().client.get_bug(webhook_private_comment_example.bug.id)
+    expanded = bugzilla_client.get_bug(webhook_private_comment_example.bug.id)
 
     # then
     assert expanded.comment.creator == "mathieu@mozilla.org"
@@ -173,11 +190,13 @@ def test_bugzilla_get_bug_comment(mocked_responses, webhook_private_comment_exam
 
 @pytest.mark.no_mocked_bugzilla
 def test_bugzilla_missing_private_comment(
+    bugzilla_client,
+    settings,
     mocked_responses,
     webhook_private_comment_example,
 ):
     bug_url = (
-        f"{get_settings().bugzilla_base_url}/rest/bug/%s"
+        f"{settings.bugzilla_base_url}/rest/bug/%s"
         % webhook_private_comment_example.bug.id
     )
     mocked_responses.add(
@@ -194,14 +213,14 @@ def test_bugzilla_missing_private_comment(
         },
     )
 
-    expanded = get_service().client.get_bug(webhook_private_comment_example.bug.id)
+    expanded = bugzilla_client.get_bug(webhook_private_comment_example.bug.id)
 
     assert not expanded.comment
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_list_webhooks(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/webhooks/list"
+def test_bugzilla_list_webhooks(bugzilla_client, settings, mocked_responses):
+    url = f"{settings.bugzilla_base_url}/rest/webhooks/list"
     mocked_responses.add(
         responses.GET,
         url,
@@ -222,7 +241,7 @@ def test_bugzilla_list_webhooks(mocked_responses):
         },
     )
 
-    webhooks = get_service().client.list_webhooks()
+    webhooks = bugzilla_client.list_webhooks()
 
     assert len(webhooks) == 1
     assert webhooks[0].event == "create,change,comment"
@@ -230,11 +249,13 @@ def test_bugzilla_list_webhooks(mocked_responses):
 
 
 @pytest.mark.no_mocked_bugzilla
-def test_bugzilla_list_webhooks_raises_if_response_has_no_webhooks(mocked_responses):
-    url = f"{get_settings().bugzilla_base_url}/rest/webhooks/list"
+def test_bugzilla_list_webhooks_raises_if_response_has_no_webhooks(
+    bugzilla_client, settings, mocked_responses
+):
+    url = f"{settings.bugzilla_base_url}/rest/webhooks/list"
     mocked_responses.add(responses.GET, url, json={})
 
     with pytest.raises(BugzillaClientError) as exc:
-        get_service().client.list_webhooks()
+        bugzilla_client.list_webhooks()
 
     assert "Unexpected response" in str(exc)
